Moving middleware out of vite config
