+++ websiteMiddlewares.ts
@@ -1,3 +1,5 @@
+import type {Plugin} from 'vite'
+
 import fs from 'fs-extra'
 import path from 'node:path'
 import slugify from 'slugify'
@@ -110,3 +112,41 @@ function createDefendersFileMap(): DefendersDataMap {
 
 export const defendersData = createDefendersData()
 export const defendersFileMap = createDefendersFileMap()
+
+const data = JSON.stringify({data: defendersData})
+const contentTypes = {
+  mp3: 'audio/mpeg',
+  mp4: 'video/mp4',
+  pdf: 'application/pdf',
+}
+
+// https://vitejs.dev/guide/api-plugin.html#configureserver
+export function devMiddleware(): Plugin {
+  return {
+    name: 'dev-middleware',
+    configureServer(server) {
+      server.middlewares.use((req, res, next) => {
+        if (req.method !== 'GET') return next()
+
+        if (req.url === '/defenders/data') {
+          return res.end(data)
+        }
+
+        for (const [fileName, filePath] of Object.entries(defendersFileMap)) {
+          if (req.url === `/defenders/${fileName}`) {
+            const {size} = fs.statSync(filePath)
+            res.writeHead(200, {
+              'Content-Type': contentTypes[filePath.slice(-3)],
+              'Content-Length': size,
+            })
+
+            const readStream = fs.createReadStream(filePath)
+            return readStream.pipe(res, {end: true})
+          }
+        }
+
+        next()
+      })
+    },
+  }
+}
